- SDK Updated. - Coded new sort option: by date. - Updated localized strings.
