+++ ownCloud/Client/SortMethod.swift
@@ -13,91 +13,101 @@ typealias OCSort = Comparator
 
 public enum SortMethod: Int {
 
-    case alphabeticallyAscendant = 0
-    case alphabeticallyDescendant = 1
-    case type = 2
-    case size = 3
+	case alphabeticallyAscendant = 0
+	case alphabeticallyDescendant = 1
+	case type = 2
+	case size = 3
+	case date = 4
 
-    static var all: [SortMethod] = [alphabeticallyAscendant, alphabeticallyDescendant, type, size]
+	static var all: [SortMethod] = [alphabeticallyAscendant, alphabeticallyDescendant, type, size, date]
 
-    func localizedName() -> String {
-        var name = ""
+	func localizedName() -> String {
+		var name = ""
 
-        switch self {
-        case .alphabeticallyAscendant:
-            name = "A-Z".localized
-        case .alphabeticallyDescendant:
-            name = "Z-A".localized
-        case .type:
-            name = "Type".localized
-        case .size:
-            name = "Size".localized
-        }
+		switch self {
+		case .alphabeticallyAscendant:
+			name = "name (A-Z)".localized
+		case .alphabeticallyDescendant:
+			name = "name (Z-A)".localized
+		case .type:
+			name = "type".localized
+		case .size:
+			name = "size".localized
+		case .date:
+			name = "date".localized
+		}
 
-        return name
-    }
+		return name
+	}
 
-    func comparator() -> OCSort? {
-        var comparator: OCSort? = nil
+	func comparator() -> OCSort {
+		var comparator: OCSort
 
-        switch self {
-        case .size:
-            comparator = { (left, right) in
-                let leftItem = left as? OCItem
-                let rightItem = right as? OCItem
+		switch self {
+		case .size:
+			comparator = { (left, right) in
+				let leftItem = left as? OCItem
+				let rightItem = right as? OCItem
 
-                let leftSize = leftItem!.size as NSNumber
-                let rightSize = rightItem!.size as NSNumber
+				let leftSize = leftItem!.size as NSNumber
+				let rightSize = rightItem!.size as NSNumber
 
-                return (rightSize.compare(leftSize))
-            }
+				return (rightSize.compare(leftSize))
+			}
 
-        case .alphabeticallyAscendant:
-            comparator = { (left, right) in
-                let leftItem = left as? OCItem
-                let rightItem = right as? OCItem
+		case .alphabeticallyAscendant:
+			comparator = { (left, right) in
+				let leftItem = left as? OCItem
+				let rightItem = right as? OCItem
 
-                return (leftItem?.name.compare(rightItem!.name))!
+				return (leftItem?.name.lowercased().compare(rightItem!.name.lowercased()))!
+			}
 
-            }
+		case .alphabeticallyDescendant:
+			comparator = {
+				(left, right) in
+				let leftItem = left as? OCItem
+				let rightItem = right as? OCItem
 
-        case .alphabeticallyDescendant:
-            comparator = {
-                (left, right) in
-                let leftItem = left as? OCItem
-                let rightItem = right as? OCItem
+				return (rightItem?.name.lowercased().compare(leftItem!.name.lowercased()))!
+			}
 
-                return (rightItem?.name.compare(leftItem!.name))!
-            }
+		case .type:
+			comparator = {
+				(left, right) in
+				let leftItem = left as? OCItem
+				let rightItem = right as? OCItem
 
-        case .type:
-            comparator = {
-                (left, right) in
-                let leftItem = left as? OCItem
-                let rightItem = right as? OCItem
+				var leftMimeType = leftItem?.mimeType
+				var rightMimeType = rightItem?.mimeType
 
-                var leftMimeType = leftItem?.mimeType
-                var rightMimeType = rightItem?.mimeType
+				if leftItem?.type == OCItemType.collection {
+					leftMimeType = "folder"
+				}
 
-                if leftItem?.type == OCItemType.collection {
-                    leftMimeType = "folder"
-                }
+				if rightItem?.type == OCItemType.collection {
+					rightMimeType = "folder"
+				}
 
-                if rightItem?.type == OCItemType.collection {
-                    rightMimeType = "folder"
-                }
+				if leftItem?.mimeType == nil {
+					leftMimeType = "various"
+				}
 
-                if leftItem?.mimeType == nil {
-                    leftMimeType = "various"
-                }
+				if rightItem?.mimeType == nil {
+					rightMimeType = "various"
+				}
 
-                if rightItem?.mimeType == nil {
-                    rightMimeType = "various"
-                }
+				return leftMimeType!.compare(rightMimeType!)
+			}
+		case .date:
+			comparator = {
+				(left, right) in
+				let leftItem = left as? OCItem
+				let rightItem = right as? OCItem
 
-                return leftMimeType!.compare(rightMimeType!)
-            }
-        }
-        return comparator
-    }
+				return (rightItem?.lastModified.compare(leftItem!.lastModified))!
+			}
+		}
+		return comparator
+	}
 }
